Fix pg_autoctl drop node with the new DROPPED state.
