+++ src/bin/pg_autoctl/cli_create_node.c
@@ -1,5 +1,5 @@
 /*
- * src/bin/pg_autoctl/cli_create_drop_node.c
+ * src/bin/pg_autoctl/cli_create_node.c
  *     Implementation of the pg_autoctl create and pg_autoctl drop CLI for the
  *     pg_auto_failover nodes (monitor, coordinator, worker, postgres).
  *
@@ -45,19 +45,12 @@
  * functions and their command implementation. We can't pass parameters around.
  */
 MonitorConfig monitorOptions = { 0 };
-static bool dropAndDestroy = false;
 
 static int cli_create_postgres_getopts(int argc, char **argv);
 static void cli_create_postgres(int argc, char **argv);
 
 static int cli_create_monitor_getopts(int argc, char **argv);
 static void cli_create_monitor(int argc, char **argv);
-
-static int cli_drop_node_getopts(int argc, char **argv);
-static void cli_drop_node(int argc, char **argv);
-static void cli_drop_monitor(int argc, char **argv);
-
-static void cli_drop_node_from_monitor(KeeperConfig *config);
 
 static void check_hostname(const char *hostname);
 
@@ -102,30 +95,6 @@ CommandLine create_postgres_command =
 		cli_create_postgres_getopts,
 		cli_create_postgres);
 
-CommandLine drop_monitor_command =
-	make_command("monitor",
-				 "Drop the pg_auto_failover monitor",
-				 "[ --pgdata --destroy ]",
-				 "  --pgdata      path to data directory\n"
-				 "  --destroy     also destroy Postgres database\n",
-				 cli_drop_node_getopts,
-				 cli_drop_monitor);
-
-CommandLine drop_node_command =
-	make_command(
-		"node",
-		"Drop a node from the pg_auto_failover monitor",
-		"[ [ [ --pgdata ] [ --destroy ] ] | "
-		"[ --monitor [ [ --hostname --pgport ] | [ --formation --name ] ] ] ] ",
-		"  --pgdata      path to data directory\n"
-		"  --monitor     pg_auto_failover Monitor Postgres URL\n"
-		"  --formation   pg_auto_failover formation\n"
-		"  --name        drop the node with the given node name\n"
-		"  --hostname    drop the node with given hostname and pgport\n"
-		"  --pgport      drop the node with given hostname and pgport\n"
-		"  --destroy     also destroy Postgres database\n",
-		cli_drop_node_getopts,
-		cli_drop_node);
 
 /*
  * cli_create_config manages the whole set of configuration parameters that
@@ -846,432 +815,6 @@ cli_create_monitor(int argc, char **argv)
 	{
 		/* errors have been logged */
 		exit(EXIT_CODE_INTERNAL_ERROR);
-	}
-}
-
-
-/*
- * cli_drop_node_getopts parses the command line options necessary to drop or
- * destroy a local pg_autoctl node.
- */
-static int
-cli_drop_node_getopts(int argc, char **argv)
-{
-	KeeperConfig options = { 0 };
-	int c, option_index = 0;
-	int verboseCount = 0;
-
-	static struct option long_options[] = {
-		{ "pgdata", required_argument, NULL, 'D' },
-		{ "monitor", required_argument, NULL, 'm' },
-		{ "destroy", no_argument, NULL, 'd' },
-		{ "hostname", required_argument, NULL, 'n' },
-		{ "pgport", required_argument, NULL, 'p' },
-		{ "formation", required_argument, NULL, 'f' },
-		{ "name", required_argument, NULL, 'a' },
-		{ "version", no_argument, NULL, 'V' },
-		{ "verbose", no_argument, NULL, 'v' },
-		{ "quiet", no_argument, NULL, 'q' },
-		{ "help", no_argument, NULL, 'h' },
-		{ NULL, 0, NULL, 0 }
-	};
-
-	optind = 0;
-
-	while ((c = getopt_long(argc, argv, "D:dn:p:Vvqh",
-							long_options, &option_index)) != -1)
-	{
-		switch (c)
-		{
-			case 'D':
-			{
-				strlcpy(options.pgSetup.pgdata, optarg, MAXPGPATH);
-				log_trace("--pgdata %s", options.pgSetup.pgdata);
-				break;
-			}
-
-			case 'm':
-			{
-				if (!validate_connection_string(optarg))
-				{
-					log_fatal("Failed to parse --monitor connection string, "
-							  "see above for details.");
-					exit(EXIT_CODE_BAD_ARGS);
-				}
-				strlcpy(options.monitor_pguri, optarg, MAXCONNINFO);
-				log_trace("--monitor %s", options.monitor_pguri);
-				break;
-			}
-
-			case 'd':
-			{
-				dropAndDestroy = true;
-				log_trace("--destroy");
-				break;
-			}
-
-			case 'n':
-			{
-				strlcpy(options.hostname, optarg, _POSIX_HOST_NAME_MAX);
-				log_trace("--hostname %s", options.hostname);
-				break;
-			}
-
-			case 'p':
-			{
-				if (!stringToInt(optarg, &options.pgSetup.pgport))
-				{
-					log_fatal("--pgport argument is a valid port number: \"%s\"",
-							  optarg);
-					exit(EXIT_CODE_BAD_ARGS);
-				}
-				log_trace("--pgport %d", options.pgSetup.pgport);
-				break;
-			}
-
-			case 'f':
-			{
-				strlcpy(options.formation, optarg, NAMEDATALEN);
-				log_trace("--formation %s", options.formation);
-				break;
-			}
-
-			case 'a':
-			{
-				/* { "name", required_argument, NULL, 'a' }, */
-				strlcpy(options.name, optarg, _POSIX_HOST_NAME_MAX);
-				log_trace("--name %s", options.name);
-				break;
-			}
-
-			case 'V':
-			{
-				/* keeper_cli_print_version prints version and exits. */
-				keeper_cli_print_version(argc, argv);
-				break;
-			}
-
-			case 'v':
-			{
-				++verboseCount;
-				switch (verboseCount)
-				{
-					case 1:
-					{
-						log_set_level(LOG_INFO);
-						break;
-					}
-
-					case 2:
-					{
-						log_set_level(LOG_DEBUG);
-						break;
-					}
-
-					default:
-					{
-						log_set_level(LOG_TRACE);
-						break;
-					}
-				}
-				break;
-			}
-
-			case 'q':
-			{
-				log_set_level(LOG_ERROR);
-				break;
-			}
-
-			case 'h':
-			{
-				commandline_help(stderr);
-				exit(EXIT_CODE_QUIT);
-				break;
-			}
-
-			default:
-			{
-				/* getopt_long already wrote an error message */
-				commandline_help(stderr);
-				exit(EXIT_CODE_BAD_ARGS);
-				break;
-			}
-		}
-	}
-
-	if (dropAndDestroy &&
-		(!IS_EMPTY_STRING_BUFFER(options.hostname) ||
-		 options.pgSetup.pgport != 0))
-	{
-		log_error("Please use either [ --hostname --pgport ] "
-				  " or [ --formation --name ] to target a remote node, "
-				  " or --destroy to destroy the local node.");
-		log_info("Destroying a node is not supported from a distance");
-		exit(EXIT_CODE_BAD_ARGS);
-	}
-
-	/* now that we have the command line parameters, prepare the options */
-	/* when we have a monitor URI we don't need PGDATA */
-	if (cli_use_monitor_option(&options))
-	{
-		if (!IS_EMPTY_STRING_BUFFER(options.pgSetup.pgdata))
-		{
-			log_warn("Given --monitor URI, the --pgdata option is ignored");
-			log_info("Connecting to monitor at \"%s\"", options.monitor_pguri);
-
-			/* the rest of the program needs pgdata actually empty */
-			bzero((void *) options.pgSetup.pgdata,
-				  sizeof(options.pgSetup.pgdata));
-		}
-	}
-	else
-	{
-		(void) prepare_keeper_options(&options);
-	}
-
-	/*
-	 * pg_autoctl drop node can be used with one of those set of arguments:
-	 *   --pgdata ...                 # to drop the local node
-	 *   --pgdata <monitor>           # to drop any node from the monitor
-	 *   --formation ... --name ...   # address a node on the monitor
-	 *   --hostname ... --pgport ...  # address a node on the monitor
-	 *
-	 * We check about the PGDATA being related to a monitor or a keeper later,
-	 * here we focus on the optargs. Remember that --formation can be skipped
-	 * to mean "default", and --pgport can be skipped to mean either PGPORT
-	 * from the environment or just 5432.
-	 */
-	if (!IS_EMPTY_STRING_BUFFER(options.name) &&
-		!IS_EMPTY_STRING_BUFFER(options.hostname))
-	{
-		log_fatal("pg_autoctl drop node target can either be specified "
-				  "using [ --formation --name ], or "
-				  "using [ --hostname and --pgport ], but not both.");
-		exit(EXIT_CODE_BAD_ARGS);
-	}
-
-	/* use the "default" formation when not given */
-	if (IS_EMPTY_STRING_BUFFER(options.formation))
-	{
-		strlcpy(options.formation, FORMATION_DEFAULT, NAMEDATALEN);
-	}
-
-	/* publish our option parsing in the global variable */
-	keeperOptions = options;
-
-	return optind;
-}
-
-
-/*
- * cli_drop_node removes the local PostgreSQL node from the pg_auto_failover
- * monitor, and when it's a worker, from the Citus coordinator too.
- */
-static void
-cli_drop_node(int argc, char **argv)
-{
-	KeeperConfig config = keeperOptions;
-
-	pgAutoCtlNodeRole localNodeRole =
-		IS_EMPTY_STRING_BUFFER(config.pgSetup.pgdata)
-		? PG_AUTOCTL_ROLE_UNKNOWN
-		: ProbeConfigurationFileRole(config.pathnames.config);
-
-	bool dropLocalNode =
-		!IS_EMPTY_STRING_BUFFER(config.pgSetup.pgdata) &&
-		localNodeRole == PG_AUTOCTL_ROLE_KEEPER;
-
-	/*
-	 * The configuration file is the last bit we remove, so we don't have to
-	 * implement "continue from previous failed attempt" when the configuration
-	 * file does not exists.
-	 */
-	if (dropLocalNode && !file_exists(config.pathnames.config))
-	{
-		log_error("Failed to find expected configuration file \"%s\"",
-				  config.pathnames.config);
-		exit(EXIT_CODE_BAD_CONFIG);
-	}
-
-	if (dropLocalNode)
-	{
-		bool missingPgdataIsOk = true;
-		bool pgIsNotRunningIsOk = true;
-		bool monitorDisabledIsOk = true;
-
-		if (!IS_EMPTY_STRING_BUFFER(config.hostname) ||
-			config.pgSetup.pgport != 0)
-		{
-			log_fatal("Only dropping the local node is supported, "
-					  "[ --hostname --pgport ] are not supported "
-					  "when --pgdata is used.");
-			log_info("To drop another node, please use this command "
-					 "from the monitor itself.");
-			exit(EXIT_CODE_BAD_ARGS);
-		}
-
-		if (!IS_EMPTY_STRING_BUFFER(config.name))
-		{
-			log_fatal("Only dropping the local node is supported, "
-					  "[ --formation --name ] are not supported "
-					  "when --pgdata is used.");
-			log_info("To drop another node, please use this command "
-					 "from the monitor itself.");
-			exit(EXIT_CODE_BAD_ARGS);
-		}
-
-		/* just read the keeper file in given KeeperConfig */
-		if (!keeper_config_read_file(&config,
-									 missingPgdataIsOk,
-									 pgIsNotRunningIsOk,
-									 monitorDisabledIsOk))
-		{
-			exit(EXIT_CODE_BAD_CONFIG);
-		}
-
-		/* drop the node and maybe destroy its PGDATA entirely. */
-		(void) cli_drop_local_node(&config, dropAndDestroy);
-
-		return;
-	}
-	else
-	{
-		/* pg_autoctl drop node on the monitor drops another node */
-		if (IS_EMPTY_STRING_BUFFER(config.name) &&
-			IS_EMPTY_STRING_BUFFER(config.hostname))
-		{
-			log_fatal("pg_autoctl drop node target can either be specified "
-					  "using [ --formation --name ], or "
-					  "using [ --hostname and --pgport ], "
-					  "please use either one.");
-			exit(EXIT_CODE_BAD_ARGS);
-		}
-
-		(void) cli_drop_node_from_monitor(&config);
-	}
-}
-
-
-/*
- * cli_drop_monitor removes the local monitor node.
- */
-static void
-cli_drop_monitor(int argc, char **argv)
-{
-	KeeperConfig config = keeperOptions;
-
-	bool missingPgdataIsOk = true;
-	bool pgIsNotRunningIsOk = true;
-
-	/*
-	 * The configuration file is the last bit we remove, so we don't have to
-	 * implement "continue from previous failed attempt" when the configuration
-	 * file does not exists.
-	 */
-	if (!file_exists(config.pathnames.config))
-	{
-		log_error("Failed to find expected configuration file \"%s\"",
-				  config.pathnames.config);
-		exit(EXIT_CODE_BAD_CONFIG);
-	}
-
-	/*
-	 * We are going to need to use the right pg_ctl binary to control the
-	 * Postgres cluster: pg_ctl stop.
-	 */
-	switch (ProbeConfigurationFileRole(config.pathnames.config))
-	{
-		case PG_AUTOCTL_ROLE_MONITOR:
-		{
-			MonitorConfig mconfig = { 0 };
-
-			if (!monitor_config_init_from_pgsetup(&mconfig,
-												  &(config.pgSetup),
-												  missingPgdataIsOk,
-												  pgIsNotRunningIsOk))
-			{
-				/* errors have already been logged */
-				exit(EXIT_CODE_BAD_CONFIG);
-			}
-
-			/* expose the pgSetup in the given KeeperConfig */
-			config.pgSetup = mconfig.pgSetup;
-
-			/* somehow at this point we've lost our pathnames */
-			if (!keeper_config_set_pathnames_from_pgdata(
-					&(config.pathnames),
-					config.pgSetup.pgdata))
-			{
-				/* errors have already been logged */
-				exit(EXIT_CODE_BAD_ARGS);
-			}
-
-			/* drop the node and maybe destroy its PGDATA entirely. */
-			(void) cli_drop_local_node(&config, dropAndDestroy);
-			return;
-		}
-
-		case PG_AUTOCTL_ROLE_KEEPER:
-		{
-			log_fatal("Local node is not a monitor");
-			exit(EXIT_CODE_BAD_CONFIG);
-
-			break;
-		}
-
-		default:
-		{
-			log_fatal("Unrecognized configuration file \"%s\"",
-					  config.pathnames.config);
-			exit(EXIT_CODE_BAD_CONFIG);
-		}
-	}
-}
-
-
-/*
- * cli_drop_node_from_monitor calls pgautofailover.remove_node() on the monitor
- * for the given --hostname and --pgport, or from the given --formation and
- * --name.
- */
-static void
-cli_drop_node_from_monitor(KeeperConfig *config)
-{
-	Monitor monitor = { 0 };
-
-	(void) cli_monitor_init_from_option_or_config(&monitor, config);
-
-	if (!IS_EMPTY_STRING_BUFFER(config->name))
-	{
-		if (!monitor_remove_by_nodename(&monitor,
-										(char *) config->formation,
-										(char *) config->name))
-		{
-			/* errors have already been logged */
-			exit(EXIT_CODE_MONITOR);
-		}
-	}
-	else if (!IS_EMPTY_STRING_BUFFER(config->hostname))
-	{
-		int pgport =
-			config->pgSetup.pgport > 0
-			? config->pgSetup.pgport
-			: pgsetup_get_pgport();
-
-		if (!monitor_remove_by_hostname(&monitor,
-										(char *) config->hostname,
-										pgport))
-		{
-			/* errors have already been logged */
-			exit(EXIT_CODE_MONITOR);
-		}
-	}
-	else
-	{
-		log_fatal("BUG: cli_drop_node_from_monitor options contain "
-				  " neither --name nor --hostname");
-		exit(EXIT_CODE_BAD_ARGS);
 	}
 }
