+++ src/bin/pg_autoctl/pidfile.c
@@ -497,3 +497,49 @@ pidfile_as_json(JSON_Value *js, const char *pidfile, bool includeStatus)
 
 	free(fileContents);
 }
+
+
+/*
+ * wait_for_pid_to_exit waits until the PID found in the pidfile is not running
+ * anymore.
+ */
+bool
+wait_for_pid_to_exit(const char *pidfile, int timeout, pid_t *pid)
+{
+	if (file_exists(pidfile))
+	{
+		if (read_pidfile(pidfile, pid))
+		{
+			log_info("An instance of pg_autoctl is running with PID %d, "
+					 "waiting for it to stop.", *pid);
+
+			int timeout_counter = timeout;
+
+			while (timeout_counter > 0)
+			{
+				if (kill(*pid, 0) == -1 && errno == ESRCH)
+				{
+					log_info("The pg_autoctl instance with pid %d "
+							 "has now terminated.",
+							 *pid);
+					break;
+				}
+
+				sleep(1);
+				--timeout_counter;
+			}
+
+			return timeout_counter > 0;
+		}
+		else
+		{
+			log_error("Failed to read PID file \"%s\"", pidfile);
+			return false;
+		}
+	}
+	else
+	{
+		/* when the pidfile doesn't exist, assume the service is not running */
+		return true;
+	}
+}
