+++ src/bin/pg_autoctl/pidfile.h
@@ -67,4 +67,6 @@ void check_pidfile(const char *pidfile, pid_t start_pid);
 
 void pidfile_as_json(JSON_Value *js, const char *pidfile, bool includeStatus);
 
+bool wait_for_pid_to_exit(const char *pidfile, int timeout, pid_t *pid);
+
 #endif /* PIDFILE_H */
